# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "家庭故事豆 API",
    "description": "家庭故事豆 MVP 的 API 契约与 Mock 联调服务。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/recordings": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "List Recordings",
        "operationId": "list_recordings_v1_recordings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_RecordingData__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Create Recording",
        "operationId": "create_recording_v1_recordings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CreateRecordingData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/uploads/{token}": {
      "put": {
        "tags": [
          "recordings"
        ],
        "summary": "Direct Upload",
        "operationId": "direct_upload_v1_uploads__token__put",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UploadReceiptData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recordings/{recording_id}/upload-complete": {
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Complete Upload",
        "operationId": "complete_upload_v1_recordings__recording_id__upload_complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recording Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CompleteUploadData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recordings/{recording_id}": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "Get Recording",
        "operationId": "get_recording_v1_recordings__recording_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RecordingData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recordings"
        ],
        "summary": "Delete Recording Draft",
        "operationId": "delete_recording_draft_v1_recordings__recording_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeleteRecordingData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/privacy/recordings/{recording_id}": {
      "delete": {
        "tags": [
          "timeline-privacy"
        ],
        "summary": "Delete Recording",
        "operationId": "delete_recording_v1_privacy_recordings__recording_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeleteResultData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "timestamp"
        ],
        "title": "ApiMeta"
      },
...
      "ApiResponse_CompleteUploadData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/CompleteUploadData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[CompleteUploadData]"
      },
      "ApiResponse_CreateRecordingData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/CreateRecordingData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[CreateRecordingData]"
      },
...
      "ApiResponse_DeleteRecordingData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/DeleteRecordingData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[DeleteRecordingData]"
      },
      "ApiResponse_DeleteResultData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/DeleteResultData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[DeleteResultData]"
      },
...
      "ApiResponse_RecordingData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/RecordingData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[RecordingData]"
      },
...
      "ApiResponse_UploadReceiptData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/UploadReceiptData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[UploadReceiptData]"
      },
...
      "ApiResponse_list_RecordingData__": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RecordingData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[RecordingData]]"
      },
...
      "CompleteUploadData": {
        "properties": {
          "recording": {
            "$ref": "#/components/schemas/RecordingData"
          },
          "job": {
            "$ref": "#/components/schemas/JobData"
          }
        },
        "type": "object",
        "required": [
          "recording",
          "job"
        ],
        "title": "CompleteUploadData"
      },
      "CompleteUploadRequest": {
        "properties": {
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "sha256"
        ],
        "title": "CompleteUploadRequest"
      },
...
      "CreateRecordingData": {
        "properties": {
          "recording": {
            "$ref": "#/components/schemas/RecordingData"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadTargetData"
          }
        },
        "type": "object",
        "required": [
          "recording",
          "upload"
        ],
        "title": "CreateRecordingData"
      },
      "CreateRecordingRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "original_file_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Original File Name"
          },
          "declared_media_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared Media Type"
          },
          "declared_size": {
            "type": "integer",
            "maximum": 524288000.0,
            "exclusiveMinimum": 0.0,
            "title": "Declared Size"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "audio_upload",
              "video_upload",
              "mobile_recording",
              "recording_bean"
            ],
            "title": "Source Type",
            "default": "audio_upload"
          }
        },
        "type": "object",
        "required": [
          "title",
          "original_file_name",
          "declared_size",
          "sha256"
        ],
        "title": "CreateRecordingRequest"
      },
...
      "DeleteRecordingData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "deleted": {
            "type": "boolean",
            "const": true,
            "title": "Deleted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DeleteRecordingData"
      },
      "DeleteResultData": {
        "properties": {
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "deleted": {
            "type": "boolean",
            "const": true,
            "title": "Deleted",
            "default": true
          },
          "object_count": {
            "type": "integer",
            "title": "Object Count"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "target_id",
          "scope",
          "object_count",
          "completed_at"
        ],
        "title": "DeleteResultData"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "stage": {
            "$ref": "#/components/schemas/JobStage"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "retry_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Retry Count"
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Error Detail"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "dead_lettered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dead Lettered At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recording_id",
          "stage",
          "progress",
          "retry_count",
          "pipeline_version"
        ],
        "title": "JobData"
      },
      "JobStage": {
        "type": "string",
        "enum": [
          "CREATED",
          "UPLOADING",
          "UPLOADED",
          "PREPROCESSING",
          "TRANSCRIBING",
          "ANALYZING",
          "READY_FOR_SELECTION",
          "GENERATING",
          "COMPLETED",
          "FAILED",
          "CANCELLED",
          "DELETED"
        ],
        "title": "JobStage"
      },
...
      "RecordingData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "family_id": {
            "type": "string",
            "format": "uuid",
            "title": "Family Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_file_name": {
            "type": "string",
            "title": "Original File Name"
          },
          "media_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Type"
          },
          "file_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "audio_upload",
              "video_upload",
              "mobile_recording",
              "recording_bean"
            ],
            "title": "Source Type",
            "default": "audio_upload"
          },
          "status": {
            "$ref": "#/components/schemas/JobStage"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "delete_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delete At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "family_id",
          "title",
          "original_file_name",
          "media_type",
          "file_size",
          "duration_ms",
          "sha256",
          "status",
          "delete_at",
          "created_at",
          "updated_at"
        ],
        "title": "RecordingData"
      },
...
      "UploadReceiptData": {
        "properties": {
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "bytes_received": {
            "type": "integer",
            "title": "Bytes Received"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "bytes_received",
          "sha256"
        ],
        "title": "UploadReceiptData"
      },
      "UploadTargetData": {
        "properties": {
          "method": {
            "type": "string",
            "const": "PUT",
            "title": "Method",
            "default": "PUT"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_at",
          "headers"
        ],
        "title": "UploadTargetData"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}